+++ crates/adapters/mock-zkvm/src/lib.rs
@@ -68,23 +68,6 @@ impl Zkvm for MockZkvm {
 )]
 pub struct MockCodeCommitment(pub [u8; 8]);
 
-impl sov_rollup_interface::zk::CodeCommitment for MockCodeCommitment {
-    type DecodeError = MockCodeCommitmentError;
-
-    fn encode(&self) -> Vec<u8> {
-        self.0.to_vec()
-    }
-
-    fn decode(value: &[u8]) -> Result<Self, Self::DecodeError> {
-        if value.len() != 8 {
-            return Err(MockCodeCommitmentError::InvalidLength { found: value.len() });
-        }
-        let mut contents = [0u8; 8];
-        contents.copy_from_slice(value);
-        Ok(Self(contents))
-    }
-}
-
 /// An error that can occur when converting a byte vector to a `MockCodeCommitment`.
 #[derive(Debug, Clone, PartialEq, Eq, Error)]
 pub enum MockCodeCommitmentError {
@@ -144,7 +127,7 @@ impl sov_rollup_interface::zk::ZkVerifier for MockZkVerifier {
 #[cfg(test)]
 mod tests {
     use sov_rollup_interface::crypto::PublicKey;
-    use sov_rollup_interface::zk::{CodeCommitment, ZkVerifier, ZkvmHost, ZkvmNetwork};
+    use sov_rollup_interface::zk::{ZkVerifier, ZkvmHost, ZkvmNetwork};
 
     use super::*;
 
@@ -242,23 +225,5 @@ mod tests {
         let verified = MockZkVerifier::verify::<TestPublicData>(&proof_bytes, &Default::default())?;
         assert_eq!(verified, pub_data);
         Ok(())
-    }
-
-    #[test]
-    fn mock_code_commitment_codec_roundtrip() {
-        // Check a roundtrip with the "digest" type from risc0.
-        // This ensures that our use of `from_ne_bytes` is correct on the target platform.
-        let raw_data = [1; 8];
-        let method_id = MockCodeCommitment(raw_data);
-        let bytes = method_id.encode();
-        let id = MockCodeCommitment::decode(&bytes).expect("Encoding is valid");
-        assert_eq!(id.0, raw_data);
-
-        // Assert that we return the expected error when the length is incorrect.
-        let bytes = vec![1u8; 31];
-        assert!(matches!(
-            MockCodeCommitment::decode(&bytes),
-            Err(MockCodeCommitmentError::InvalidLength { found: 31 })
-        ));
     }
 }