+++ examples/demo-rollup/tests/prover/mod.rs
@@ -16,8 +16,7 @@ use sov_rollup_interface::node::da::DaService;
 use sov_rollup_interface::stf::{ExecutionContext, StateTransitionFunction};
 use sov_rollup_interface::storage::HierarchicalStorageManager;
 use sov_rollup_interface::zk::{
-    CodeCommitment, StateTransitionPublicData, StateTransitionWitness,
-    StateTransitionWitnessWithAddress, ZkvmHost,
+    StateTransitionPublicData, StateTransitionWitness, StateTransitionWitnessWithAddress, ZkvmHost,
 };
 use sov_sp1_adapter::host::SP1Host;
 use sov_sp1_adapter::BlockHeaderWithProof;
@@ -220,7 +219,7 @@ impl TestHost {
     }
 
     fn verifying_key_bytes(&self) -> Vec<u8> {
-        self.code_commitment.encode()
+        self.code_commitment.0.clone()
     }
 
     #[allow(dead_code)]