+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 # 2025-03-16
 # 2026-03-24
 - #2621 EVM: Fix feeHistory endpoint for zero blocks
+- #2634 Test only changes
+- #2631 Reorganize and extend demo-stf
 
 # 2026-03-16
 - #2592 Fixes EVM RPC regression for paymaster enabled rollups